fixed a bug where the doubles got sent in a wrong format

+++ phantom_module/UnityListener.h
@@ -30,7 +30,7 @@ class UnityListener
 	int connectSocketRecv, slen_recv, iResultRecv;
 	WSADATA wsaDataRecv;
 	char *SERVER;
-	int RECEIVER_PORT;
+	unsigned short RECEIVER_PORT;
 
 public:
 	UnityListener();
